Remove int to size_t comparison in ft_strncmp

+++ ft_strncmp.c
@@ -11,10 +11,11 @@
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdio.h>
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (!s1 || !s2)
 		return (0);
